Moved followers to their own app. Still need to update frontend
