+++ backend/blogs/admin.py
@@ -4,7 +4,6 @@ from blogs.models import (
     BlogPost,
     PostComment,
     CommentReaction,
-    Follower,
     PostReaction,
     Crosspost,
 )
@@ -13,5 +12,4 @@ admin.site.register(BlogPost)
 admin.site.register(PostReaction)
 admin.site.register(PostComment)
 admin.site.register(CommentReaction)
-admin.site.register(Follower)
 admin.site.register(Crosspost)
