+++ backend/followers/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+from .models import Follower
+
+admin.site.register(Follower)
